Add project import/download

+++ app/src/main/java/libre/sampler/utils/ProgressFraction.java
@@ -1,6 +1,6 @@
 package libre.sampler.utils;
 
 public interface ProgressFraction {
-    void setProgressTotal(int total);
-    void setProgressCurrent(int current);
+    void setProgressTotal(float total);
+    void setProgressCurrent(float current);
 }
